@@ -3,7 +3,7 @@
     "type": "project",
     "license": "proprietary",
     "require": {
-        "php": ">=7.2.5",
+        "php": ">=7.4.0",
         "ext-ctype": "*",
         "ext-iconv": "*",
         "api-platform/core": "^2.5",
@@ -32,6 +32,7 @@
     },
     "require-dev": {
         "doctrine/doctrine-fixtures-bundle": "^3.4",
+        "php-coveralls/php-coveralls": "^2.4",
         "phpunit/php-code-coverage": "^9.2",
         "symfony/browser-kit": "5.2.*",
         "symfony/http-client": "5.2.*",
