@@ -12,6 +12,7 @@ use App\Validator\IsUserOwnerClass;
 use Doctrine\ORM\Mapping as ORM;
 use ApiPlatform\Core\Bridge\Doctrine\Orm\Filter\DateFilter;
 use ApiPlatform\Core\Bridge\Doctrine\Orm\Filter\SearchFilter;
+use ApiPlatform\Core\Bridge\Doctrine\Orm\Filter\OrderFilter;
 use DateTimeInterface;
 use Symfony\Component\Serializer\Annotation\Groups;
 use Symfony\Component\Validator\Constraints as Assert;
@@ -21,18 +22,20 @@ use Symfony\Component\Validator\Constraints as Assert;
  * @ApiResource(
  *     collectionOperations={
  *          "get" = { "security" = "is_granted('ROLE_USER')" },
- *          "post" = {
- *              "security" = "is_granted('ROLE_USER')"
- *          },
+ *          "post" = { "security" = "is_granted('ROLE_USER')" }
  *     },
  *     itemOperations={
  *          "get" = { "security" = "is_granted('ROLE_USER') and object.getUser() == user" },
  *          "put" = { "security" = "is_granted('ROLE_USER') and object.getUser() == user" },
  *          "delete" = { "security" = "is_granted('ROLE_USER') and object.getUser() == user" }
  *     },
+ *     attributes={
+ *      "order"={"id": "DESC"}
+ *     }
  * )
  * @ApiFilter(DateFilter::class, properties={"activityDate"})
  * @ApiFilter(SearchFilter::class, properties={"id": "exact", "performendTime": "exact", "description": "partial"})
+ * @ApiFilter(OrderFilter::class, properties={"id", "performendTime", "description", "activityDate"}, arguments={"orderParameterName"="order"})
  * @IsUserOwnerClass()
  */
 class Activity
@@ -42,20 +45,20 @@ class Activity
      * @ORM\GeneratedValue
      * @ORM\Column(type="integer")
      */
-    private $id;
+    private int $id;
 
     /**
      * @ORM\Column(type="date")
      * @Groups({"user:read", "user:write"})
      */
-    private $activityDate;
+    private DateTimeInterface $activityDate;
 
     /**
      * @ORM\Column(type="float")
      * @Groups({"user:read", "user:write"})
      * @Assert\NotBlank()
      */
-    private $performendTime;
+    private float $performendTime;
 
     /**
      * @ORM\Column(type="text")
@@ -63,7 +66,7 @@ class Activity
      * @Assert\NotBlank()
      * @HappyCoder()
      */
-    private $description;
+    private string $description;
 
     /**
      * @ORM\ManyToOne(targetEntity=User::class, inversedBy="activities")
