@@ -9,13 +9,14 @@ class UserCreateTest extends CustomApiTestCase
     {
         $client = self::createClient();
 
-        $client->request('POST', '/api/users', [
+        $res = $client->request('POST', '/api/users', [
            'json' => [
                'email' => uniqid() . '@test.com',
                'password' => 'secret'
            ]
         ]);
 
+        $this->assertResponseHeaderSame('Content-Type', 'application/ld+json; charset=utf-8');
         $this->assertResponseStatusCodeSame(self::RESOURSE_CREATED_201);
     }
 }
